requirements.txt
flask==3.1.0
tensorflow==2.16.2
numpy==1.26.4
Pillow==10.4.0
redis==5.0.8
gunicorn==22.0.0
Werkzeug==3.1.3
